Revert "feat(hogql): Add `hogql` JS template tag for HogQL formatting (#16393)"
This reverts commit ed07aacb1e2c59470649c7cb664815d478eaa617.
