+++ frontend/src/lib/components/Cards/InsightCard/InsightDetails.tsx
@@ -19,8 +19,7 @@ import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
 import { Lettermark } from 'lib/lemon-ui/Lettermark'
 import { Link } from 'lib/lemon-ui/Link'
 import { ProfilePicture } from 'lib/lemon-ui/ProfilePicture'
-import { PropertyKeyInfo } from '../../PropertyKeyInfo'
-import { KEY_MAPPING } from 'lib/taxonomy'
+import { KEY_MAPPING, PropertyKeyInfo } from '../../PropertyKeyInfo'
 import { TZLabel } from '../../TZLabel'
 import { propertyDefinitionsModel } from '~/models/propertyDefinitionsModel'
 import { cohortsModel } from '~/models/cohortsModel'
