+++ frontend/src/lib/components/DefinitionPopover/DefinitionPopover.tsx
@@ -3,7 +3,7 @@ import clsx from 'clsx'
 import { definitionPopoverLogic, DefinitionPopoverState } from 'lib/components/DefinitionPopover/definitionPopoverLogic'
 import { useActions, useValues } from 'kea'
 import { TaxonomicFilterGroupType } from 'lib/components/TaxonomicFilter/types'
-import { getKeyMapping } from 'lib/taxonomy'
+import { getKeyMapping } from 'lib/components/PropertyKeyInfo'
 import { KeyMapping, UserBasicType, PropertyDefinition } from '~/types'
 import { Owner } from 'scenes/events/Owner'
 import { dayjs } from 'lib/dayjs'
