+++ frontend/src/lib/components/PropertyFilters/components/PropertyFilterButton.tsx
@@ -6,7 +6,7 @@ import { cohortsModel } from '~/models/cohortsModel'
 import { useValues } from 'kea'
 import { propertyDefinitionsModel } from '~/models/propertyDefinitionsModel'
 import { formatPropertyLabel, midEllipsis } from 'lib/utils'
-import { KEY_MAPPING } from 'lib/taxonomy'
+import { KEY_MAPPING } from 'lib/components/PropertyKeyInfo'
 import React from 'react'
 import { PropertyFilterIcon } from 'lib/components/PropertyFilters/components/PropertyFilterIcon'
