+++ frontend/src/scenes/data-management/definition/DefinitionEdit.tsx
@@ -6,7 +6,7 @@ import { LemonButton } from 'lib/lemon-ui/LemonButton'
 import { Field } from 'lib/forms/Field'
 import { LemonTextArea } from 'lib/lemon-ui/LemonTextArea/LemonTextArea'
 import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
-import { getPropertyLabel, isPostHogProp } from 'lib/taxonomy'
+import { getPropertyLabel, isPostHogProp } from 'lib/components/PropertyKeyInfo'
 import { VerifiedDefinitionCheckbox } from 'lib/components/DefinitionPopover/DefinitionPopoverContents'
 import { LemonSelect } from 'lib/lemon-ui/LemonSelect'
 import { Form } from 'kea-forms'
