+++ frontend/src/scenes/data-management/definition/definitionLogic.ts
@@ -6,7 +6,7 @@ import { updatePropertyDefinitions } from '~/models/propertyDefinitionsModel'
 import { router } from 'kea-router'
 import { urls } from 'scenes/urls'
 import type { definitionLogicType } from './definitionLogicType'
-import { getPropertyLabel } from 'lib/taxonomy'
+import { getPropertyLabel } from 'lib/components/PropertyKeyInfo'
 import { userLogic } from 'scenes/userLogic'
 import { eventDefinitionsTableLogic } from '../events/eventDefinitionsTableLogic'
 import { propertyDefinitionsTableLogic } from '../properties/propertyDefinitionsTableLogic'
