+++ frontend/src/scenes/data-management/events/eventDefinitionsTableLogic.ts
@@ -2,7 +2,7 @@ import { actions, kea, key, listeners, path, props, reducers, selectors } from '
 import { AnyPropertyFilter, Breadcrumb, EventDefinitionType, EventDefinition, PropertyDefinition } from '~/types'
 import type { eventDefinitionsTableLogicType } from './eventDefinitionsTableLogicType'
 import api, { PaginatedResponse } from 'lib/api'
-import { keyMappingKeys } from 'lib/taxonomy'
+import { keyMappingKeys } from 'lib/components/PropertyKeyInfo'
 import { actionToUrl, combineUrl, router, urlToAction } from 'kea-router'
 import { convertPropertyGroupToProperties, objectsEqual } from 'lib/utils'
 import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
