+++ frontend/src/scenes/insights/summarizeInsight.ts
@@ -1,6 +1,6 @@
 import { AnyPartialFilterType, EntityFilter, FilterType, FunnelVizType, StepOrderValue } from '~/types'
 import { BreakdownFilter, InsightQueryNode, Node, StickinessQuery } from '~/queries/schema'
-import { KEY_MAPPING } from 'lib/taxonomy'
+import { KEY_MAPPING } from 'lib/components/PropertyKeyInfo'
 import { toLocalFilters } from 'scenes/insights/filters/ActionFilter/entityFilterLogic'
 import {
     isFunnelsFilter,
