+++ frontend/src/scenes/session-recordings/playlist/playlistUtils.ts
@@ -3,7 +3,7 @@ import { cohortsModelType } from '~/models/cohortsModelType'
 import { toLocalFilters } from 'scenes/insights/filters/ActionFilter/entityFilterLogic'
 import { getDisplayNameFromEntityFilter } from 'scenes/insights/utils'
 import { convertPropertyGroupToProperties, deleteWithUndo, genericOperatorMap } from 'lib/utils'
-import { getKeyMapping } from 'lib/taxonomy'
+import { getKeyMapping } from 'lib/components/PropertyKeyInfo'
 import api from 'lib/api'
 import { lemonToast } from 'lib/lemon-ui/lemonToast'
 import { DEFAULT_RECORDING_FILTERS } from 'scenes/session-recordings/playlist/sessionRecordingsListLogic'
